chore: Move SDL2 include and lib to project root
I saw this other project called brick-breaker on github that followed
this directory structure and felt it was a more structured way of doing
things compared to having them under src. In hindsight, it was silly to
have them under src in the first place. You learn something new each
day :D

Link: https://github.com/joshuacrotts/brick-breaker